+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ArangoDB Starter Changelog
 
 ## [master](https://github.com/arangodb-helper/arangodb/tree/master) (N/A)
+- Fix upgrade with ArangoSync component enabled
+- Bump Go version (1.21.8) and dependencies for CVE fixes
 
 ## [v0.18.3](https://github.com/arangodb-helper/arangodb/tree/0.18.3) (2024-03-05)
 - Mark ArangoSync component as deprecated